Introduce 'strict protocol' mode
git-as-svn no longer silently skips unsupported command options and instead raises
exception if client sends unexpected data. This is required to make sure we're 100%
network-compatible with native Subversion and do not silently ignore important data.
In case of unexpected bugs, this mode can be disabled with `strictProtocol: false`
in git-as-svn.conf